factor out query resolvers
